+++ jkube-kit/config/image/src/test/java/org/eclipse/jkube/kit/config/image/ImageNameTest.java
@@ -17,6 +17,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.MethodSource;
 
 import java.util.stream.Stream;
@@ -159,30 +160,33 @@ class ImageNameTest {
         String name = "roman.gordill/customer-service-cache:latest";
 
         // When
-        ImageName imageName = new ImageName(name);
+        String nameWithRegistry = new ImageName(name).getFullName("quay.io");
+        ImageName imageName = new ImageName(nameWithRegistry);
 
         // Then
         assertThat(imageName).isNotNull();
         assertThat(imageName.getUser()).isEqualTo("roman.gordill");
         assertThat(imageName.getRepository()).isEqualTo("roman.gordill/customer-service-cache");
         assertThat(imageName.getTag()).isEqualTo("latest");
-        assertThat(imageName.getRegistry()).isNull();
+        assertThat(imageName.getRegistry()).isEqualTo("quay.io");
     }
 
-    @Test
-    void testImageNameWithRegistry() {
+    @ParameterizedTest
+    @CsvSource({
+        "foo.com/customer-service-cache:latest,foo.com,customer-service-cache,latest",
+        "myregistry.127.0.0.1.nip.io/eclipse-temurin:11.0.18_10-jdk-alpine,myregistry.127.0.0.1.nip.io,eclipse-temurin,11.0.18_10-jdk-alpine"
+    })
+    void testImageNameWithRegistry(String name, String registry, String repository, String tag) {
         // Given
-        String name = "foo.com/customer-service-cache:latest";
-
         // When
         ImageName imageName = new ImageName(name);
 
         // Then
         assertThat(imageName).isNotNull();
         assertThat(imageName.getUser()).isNull();
-        assertThat(imageName.getRepository()).isEqualTo("customer-service-cache");
-        assertThat(imageName.getTag()).isEqualTo("latest");
-        assertThat(imageName.getRegistry()).isEqualTo("foo.com");
+        assertThat(imageName.getRegistry()).isEqualTo(registry);
+        assertThat(imageName.getRepository()).isEqualTo(repository);
+        assertThat(imageName.getTag()).isEqualTo(tag);
     }
 
     @Test